tests: update for bug in EELS.
